zoo_adk_agent/pyproject.toml
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "zoo_adk_agent"
version = "0.1.0"
description = "A simple agent to demonstrate the ADK."
authors = [{ name = "Casey West", email = "caseywest@google.com" }]
requires-python = ">=3.13"
license = { text = "Apache-2.0" }
dependencies = [
    "google-adk",
    "fastapi",
    "uvicorn[standard]",
    "google-auth",
]
